requirements.txt
Flask==3.1.2
flask-cors==6.0.1
gunicorn==21.2.0
mysql-connector-python==9.4.0

